servicio_pacientes/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<3.16
requests>=2.28,<2.33
